EM-3140: Code cleaned up
